Move getCloudPlatformRequiredForImport to import-requirements and rename
