+++ src/import-data/import-utils.ts
@@ -1,33 +1,7 @@
-import { Snapshot } from 'src/libs/ajax/DataRepo';
-import { isFeaturePreviewEnabled } from 'src/libs/feature-previews';
-import { ENABLE_AZURE_PFB_IMPORT } from 'src/libs/feature-previews-config';
-import {
-  canWrite,
-  CloudProvider,
-  getCloudProviderFromWorkspace,
-  isProtectedWorkspace,
-  WorkspaceWrapper,
-} from 'src/workspaces/utils';
+import { canWrite, getCloudProviderFromWorkspace, isProtectedWorkspace, WorkspaceWrapper } from 'src/workspaces/utils';
 
-import { isProtectedSource } from './import-requirements';
+import { getRequiredCloudPlatform, isProtectedSource } from './import-requirements';
 import { ImportRequest } from './import-types';
-
-export const getCloudPlatformRequiredForImport = (importRequest: ImportRequest): CloudProvider | undefined => {
-  switch (importRequest.type) {
-    case 'tdr-snapshot-export':
-    case 'tdr-snapshot-reference':
-      const tdrCloudPlatformToCloudProvider: Record<Snapshot['cloudPlatform'], CloudProvider> = {
-        azure: 'AZURE',
-        gcp: 'GCP',
-      };
-      return tdrCloudPlatformToCloudProvider[importRequest.snapshot.cloudPlatform];
-    case 'pfb':
-      // restrict PFB imports to GCP unless the user has the right feature flag enabled
-      return isFeaturePreviewEnabled(ENABLE_AZURE_PFB_IMPORT) ? undefined : 'GCP';
-    default:
-      return undefined;
-  }
-};
 
 export type ImportOptions = {
   /** Authorization domain requested for destination workspace. */
@@ -48,7 +22,7 @@ export const buildDestinationWorkspaceFilter = (
   const { requiredAuthorizationDomain } = importOptions;
 
   const isProtectedData = isProtectedSource(importRequest);
-  const requiredCloudPlatform = getCloudPlatformRequiredForImport(importRequest);
+  const requiredCloudPlatform = getRequiredCloudPlatform(importRequest);
 
   return (workspace: WorkspaceWrapper): boolean => {
     // The user must be able to write to the workspace to import data.
